requirements.txt
django
django-netfields
paramiko==2.0.2
ldap3
passlib
hvac
djangorestframework
markdown
django-filter